+++ NEWS.md
@@ -1,6 +1,7 @@
 # dreamer 3.0.0.9000
 
 * Updated README.Rmd and vignettes (#1).
+* Restore existing rjags modules on exit from `dreamer_mcmc()` (#4).
 
 # dreamer 3.0.0